+++ onsiteprint-plugin/CHANGELOG.md
@@ -139,6 +139,7 @@ This document includes the change log notes for the OnsitePrint Plugin - [GitHub
     ---
 
 - Added New `Download Files Template Part`.
+- Changed the Modal variable.
 
     [EB | download-files-template.php](blocks/event/block-template-parts/templates/download-files-template.php)
 
@@ -146,8 +147,8 @@ This document includes the change log notes for the OnsitePrint Plugin - [GitHub
 
 - Added new `Reload Button` in the modal.php.
 - Added new `Cancel Error Button` and moved `Close Button`.
-- Changed the Modal variable.
 - Added new `Download Files Template Part`.
+- Changed the Modal variable.
 
     [EB | modal.php](blocks/event/block-template-parts/modal.php)